+++ SkillBladeWaltz/net/swagserv/andrew2060/heroes/skills/SkillBladeWaltz.java
@@ -82,7 +82,7 @@ public class SkillBladeWaltz extends ActiveSkill {
 	}
 
 	public class BleedSkillEffect extends PeriodicDamageEffect {
-		public BleedSkillEffect(Skill skill, long duration, long period, int tickDamage, Player applier) { 	
+		public BleedSkillEffect(Skill skill, long duration, long period, double tickDamage, Player applier) { 	
 			super(skill, "Bleed", period, duration, tickDamage, applier);
 			this.types.add(EffectType.BLEED); 
 		}
