+++ SkillCamouflage/net/swagserv/andrew2060/heroes/skills/SkillCamouflage.java
@@ -366,7 +366,7 @@ public class SkillCamouflage extends PassiveSkill {
 			}
 			cT.removeEffect(cT.getEffect("CamouflageDmg"));
 			int bonusdmg = SkillConfigManager.getUseSetting((Hero)cT, skill, "bonuspercent", 50, false);
-			Skill.damageEntity((LivingEntity)event.getEntity(), cT.getEntity(), (int) (bonusdmg*0.01*event.getDamage()), DamageCause.CUSTOM);
+			Skill.damageEntity((LivingEntity)event.getEntity(), cT.getEntity(), bonusdmg*0.01*event.getDamage(), DamageCause.CUSTOM);
 			//And we're done!
 		}
 	}
