+++ SkillEarthenArmor/net/swagserv/andrew2060/heroes/skills/SkillEarthenArmor.java
@@ -57,7 +57,7 @@ public class SkillEarthenArmor extends ActiveSkill {
 		int duration = SkillConfigManager.getUseSetting(hero, this, SkillSetting.DURATION, 300000, false);
 		List<Entity> entitylist = hero.getEntity().getNearbyEntities(10.0D, 10.0D, 10.0D);
 		Player p = hero.getPlayer();
-		int damage = p.getHealth() * 2;
+		double damage = p.getHealth() * 2;
 		for (Entity entity : entitylist) {
 			if (!(entity instanceof Player)) {
 				if (((entity instanceof LivingEntity)) && (!(entity instanceof Player))) {
