+++ SkillShadowStrike/net/swagserv/andrew2060/heroes/skills/SkillShadowStrike.java
@@ -55,7 +55,7 @@ public class SkillShadowStrike extends TargettedSkill {
 		h.getPlayer().teleport(lE.getLocation(), TeleportCause.UNKNOWN);
 		cT.addEffect(new ExpirableEffect(this, plugin, "ShadowStrikeSilence", 4000));
 		lE.addPotionEffect(PotionEffectType.SLOW.createEffect(60, 2));
-		Skill.damageEntity(lE, h.getEntity(), (int) (lE.getMaxHealth()*5*0.01), DamageCause.ENTITY_ATTACK);
+		Skill.damageEntity(lE, h.getEntity(), lE.getMaxHealth()*5*0.01, DamageCause.ENTITY_ATTACK);
 		if(!(lE instanceof Player)) {
 			broadcast(h.getPlayer().getLocation(), "§7[§2Skill§7]$1 struck $2 from the shadows.", new Object[] {h.getName(),lE.getType().getName()});
 		} else {
